+++ webapp/components/description/actions.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import * as autosave from '../autosave/actions'
+import * as autosave from '../../autosave/actions'
 import { applicationError } from '@webapp/loggedin/applicationError/actions'
 
 export const descriptionsFetched = 'descriptions/fetched'
