+++ webapp/components/description/generalComments.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import CommentableDescription from '@webapp/description/commentableDescription'
+import CommentableDescription from '@webapp/components/description/commentableDescription'
 import { connect } from 'react-redux'
 import * as UserState from '@webapp/user/userState'
