+++ webapp/components/description/style.less
@@ -1,5 +1,5 @@
-@import (reference) "../app-styles/config";
-@import (reference) "../app-styles/buttons";
+@import (reference) "../../app-styles/config";
+@import (reference) "../../app-styles/buttons";
 
 .fra-description__container {
   border-bottom: @ui-thin-border;
