+++ webapp/loggedin/appViewExport.js
@@ -7,7 +7,7 @@ import originalDataPoint from '../originalDataPoint/reducers'
 import autoSave from '../autosave/reducer'
 import navigation from './navigation/reducers'
 import traditionalTable from '@webapp/traditionalTable/reducer'
-import descriptions from '@webapp/description/reducers'
+import descriptions from '@webapp/components/description/reducers'
 import review from './review/reducer'
 import userManagement from '../userManagement/reducer'
 import country from '../country/reducer'
