+++ webapp/loggedin/assessmentFra/biomassStock/biomassStockView.js
@@ -6,9 +6,9 @@ import TraditionalTable from '@webapp/traditionalTable/traditionalTable'
 import tableSpec from './tableSpec'
 import DefinitionLink from '@webapp/components/definitionLink'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/audit/actions'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/descriptionBundles/analysisDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 import { isFRA2020SectionEditDisabled } from '@webapp/utils/assessmentAccess'
 import * as table from '@webapp/traditionalTable/table'
 import { isPrintingOnlyTables } from '@webapp/loggedin/printAssessment/printAssessment'
