+++ webapp/loggedin/assessmentFra/carbonStock/carbonStockView.js
@@ -4,9 +4,9 @@ import * as R from 'ramda'
 
 import TraditionalTable from '@webapp/traditionalTable/traditionalTable'
 import tableSpec from './tableSpec'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/descriptionBundles/analysisDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 import DefinitionLink from '@webapp/components/definitionLink'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/audit/actions'
 import { isFRA2020SectionEditDisabled } from '@webapp/utils/assessmentAccess'
