+++ webapp/loggedin/assessmentFra/extentOfForest/extentOfForestView.js
@@ -6,9 +6,9 @@ import Icon from '@webapp/components/icon'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ChartWrapper from './chart/chartWrapper'
 import { GenerateFraValuesControl, TableWithOdp } from '../../../tableWithOdp/tableWithOdp'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/descriptionBundles/analysisDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 import ReviewIndicator from '@webapp/loggedin/review/reviewIndicator'
 import TraditionalTable from '@webapp/traditionalTable/traditionalTable'
 import NationalDataPointsPrintView from '../../../originalDataPoint/nationalDataPointsPrintView'
