+++ webapp/loggedin/assessmentFra/forestCharacteristics/forestCharacteristicsView.js
@@ -6,9 +6,9 @@ import * as R from 'ramda'
 import { fetchItem, generateFraValues, save, saveMany } from '../../../tableWithOdp/actions'
 import { GenerateFraValuesControl, TableWithOdp } from '../../../tableWithOdp/tableWithOdp'
 import ChartWrapper from '../extentOfForest/chart/chartWrapper'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/descriptionBundles/analysisDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/audit/actions'
 import { saveCountryConfigSetting } from '../../../country/actions'
 import DefinitionLink from '@webapp/components/definitionLink'
