+++ webapp/loggedin/assessmentFra/graduationOfStudents/graduationOfStudentsPrintView.js
@@ -4,8 +4,8 @@ import { connect } from 'react-redux'
 import tableSpec, { tableProps } from './tableSpec'
 
 import TraditionalTable from '@webapp/traditionalTable/traditionalTable'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 
 import * as UserState from '@webapp/user/userState'
