+++ webapp/loggedin/assessmentFra/growingStock/growingStockView.js
@@ -13,9 +13,9 @@ import ButtonTableExport from '@webapp/components/buttonTableExport'
 import { ThousandSeparatedDecimalInput } from '@webapp/components/thousandSeparatedDecimalInput'
 import DefinitionLink from '@webapp/components/definitionLink'
 import ReviewIndicator from '@webapp/loggedin/review/reviewIndicator'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import AnalysisDescriptions from '@webapp/descriptionBundles/analysisDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import AnalysisDescriptions from '@webapp/components/description/analysisDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 
 import defaultYears from '@server/eof/defaultYears'
 import { isFRA2020SectionEditDisabled } from '@webapp/utils/assessmentAccess'
