+++ webapp/loggedin/assessmentFra/nonWoodForestProductsRemovals/nonWoodForestProductsRemovalsView.js
@@ -8,8 +8,8 @@ import TraditionalTable from '@webapp/traditionalTable/traditionalTable'
 import mainTableSpec, { sectionName } from './mainTableSpec'
 import DefinitionLink from '@webapp/components/definitionLink'
 import { fetchLastSectionUpdateTimestamp } from '@webapp/audit/actions'
-import NationalDataDescriptions from '@webapp/descriptionBundles/nationalDataDescriptions'
-import GeneralComments from '@webapp/descriptionBundles/generalComments'
+import NationalDataDescriptions from '@webapp/components/description/nationalDataDescriptions'
+import GeneralComments from '@webapp/components/description/generalComments'
 import { isFRA2020SectionEditDisabled } from '@webapp/utils/assessmentAccess'
 import { fetchTableData } from '@webapp/traditionalTable/actions'
 import * as table from '@webapp/traditionalTable/table'
